+++ R/boilerplate-glmnet.R
@@ -161,8 +161,8 @@ hai_auto_glmnet <- function(.data, .rec_obj, .splits_obj = NULL, .rsamp_obj = NU
     }
 
     # Tune the workflow
-    # Start parallel backed
-    modeltime::parallel_start(num_cores)
+    # Start parallel backend
+    cl <- parallelly::makeClusterPSOCK(workers = num_cores)
 
     tuned_results <- wflw %>%
       tune::tune_grid(
@@ -171,7 +171,7 @@ hai_auto_glmnet <- function(.data, .rec_obj, .splits_obj = NULL, .rsamp_obj = NU
         metrics   = ms
       )
 
-    modeltime::parallel_stop()
+    parallel::stopCluster(cl)
 
     # Get the best result set by a specified metric
     best_result_set <- tuned_results %>%