+++ R/data-prep-c50.R
@@ -41,7 +41,7 @@
 hai_c50_data_prepper <- function(.data, .recipe_formula) {
 
   # Recipe ---
-  rec_obj <- recipes::recipe(.recipe_formula, data = .data) %>%
+  rec_obj <- recipes::recipe(.recipe_formula, data = .data) |>
     recipes::step_string2factor(tidyselect::vars_select_helpers$where(is.character))
 
   # Return ----