+++ R/hai-data-scale.R
@@ -77,7 +77,7 @@
 #'   .recipe_object = rec_obj,
 #'   value,
 #'   .type_of_scale = "center"
-#' )$new_rec_obj %>%
+#' )$new_rec_obj |>
 #'   get_juiced_data()
 #'
 #' @return