+++ R/hai-data-trig.R
@@ -46,7 +46,7 @@
 #'   .recipe_object = rec_obj,
 #'   value,
 #'   .type_of_scale = "sinh"
-#' )$new_rec_obj %>%
+#' )$new_rec_obj |>
 #'   get_juiced_data()
 #'
 #' @return