+++ R/plot-hai-density.R
@@ -24,7 +24,7 @@
 #' @examples
 #' library(dplyr)
 #'
-#' df <- hai_scale_zero_one_vec(.x = mtcars$mpg) %>%
+#' df <- hai_scale_zero_one_vec(.x = mtcars$mpg) |>
 #'   hai_distribution_comparison_tbl()
 #'
 #' tidy_density_tbl <- hai_get_density_data_tbl(df)
@@ -80,8 +80,8 @@ hai_density_plot <- function(.data, .dist_name_col, .x_col, .y_col,
   }
 
   # Data setup ----
-  data_tbl <- tibble::as_tibble(.data) %>%
-    dplyr::ungroup() %>%
+  data_tbl <- tibble::as_tibble(.data) |>
+    dplyr::ungroup() |>
     dplyr::select(
       {{ dist_name_var }},
       {{ x_col_var }},