+++ NEWS.md
@@ -10,6 +10,7 @@ None
 
 ## Minor Fixes and Improvements
 - Fix #349 - Update `hai_c50_data_prepper()` from @EmilHvitfeldt
+
 # healthyR.ai 0.1.0
 
 ## Breaking Changes