+++ NEWS.md
@@ -1,3 +1,5 @@
+# healthyR.ai (development version)
+
 # healthyR.ai 0.1.1
 
 ## Breaking Changes
@@ -8,6 +10,7 @@ None
 
 ## Minor Fixes and Improvements
 - Fix #349 - Update `hai_c50_data_prepper()` from @EmilHvitfeldt
+
 # healthyR.ai 0.1.0
 
 ## Breaking Changes